+++ src/order.cpp
@@ -2,4 +2,4 @@
 
 Order::Order() : price(0), size(0), dir(NONE) {}
 
-Order::Order(int price, int size, direction dir) : price(price), size(size), dir(dir) {}
+Order::Order(double price, int size, direction dir) : price(price), size(size), dir(dir) {}
